requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.34
langchain_openai==0.1.22
numpy==1.24.3
pandas==2.0.3
Pillow==10.4.0
plotly==5.9.0
python-dotenv==1.0.1
sql_formatter==0.6.2
SQLAlchemy==1.4.39
sqlvalidator==0.0.20
streamlit==1.36.0
pymysql==1.1.1
langdetect
